only unlink if it's our own socket
previously we wer unlinking any unix socket passed via systemd. Instead we are now letting this job to systemd.

+++ gunicorn/sock.py
@@ -65,7 +65,7 @@ class BaseSocket(object):
         self.sock = None
 
     def unlink(self):
-        return True
+        return
 
 
 class TCPSocket(BaseSocket):
@@ -99,7 +99,7 @@ class UnixSocket(BaseSocket):
 
     FAMILY = socket.AF_UNIX
 
-    def __init__(self, addr, conf, log, fd=None):
+    def __init__(self, addr, conf, log, fd=None, unlink=True):
         if fd is None:
             try:
                 st = os.stat(addr)
@@ -111,6 +111,7 @@ class UnixSocket(BaseSocket):
                     os.remove(addr)
                 else:
                     raise ValueError("%r is not a socket" % addr)
+        self.can_unlink = unlink
         super(UnixSocket, self).__init__(addr, conf, log, fd=fd)
 
     def __str__(self):
@@ -126,6 +127,8 @@ class UnixSocket(BaseSocket):
         super(UnixSocket, self).close()
 
     def unlink(self):
+        if not self.can_unlink:
+            return
         os.unlink(self.cfg_addr)
 
 
@@ -163,7 +166,8 @@ def create_sockets(conf, log):
                 sock = socket.fromfd(fd, socket.AF_UNIX, socket.SOCK_STREAM)
                 sockname = sock.getsockname()
                 if isinstance(sockname, str) and sockname.startswith('/'):
-                    listeners.append(UnixSocket(sockname, conf, log, fd=fd))
+                    listeners.append(UnixSocket(sockname, conf, log, fd=fd,
+                                                unlink=False))
                 elif len(sockname) == 2 and '.' in sockname[0]:
                     listeners.append(TCPSocket("%s:%s" % sockname, conf, log,
                         fd=fd))
